Refacto : moved log page related functions mhvtl_log.c : - update_<log_page> - get/set_tape_load_status - set_current_state - set_lp11_<flag>
