+++ usr/cmd/vtltape.c
@@ -324,76 +324,6 @@ int lookup_mode_media_type(struct name_to_media_info *media_info, int med) {
 	return media_type_unknown;
 }
 
-/* Update WriteProtect bit in LogPage 0x11 */
-static void set_lp_11_wp(int flag) {
-	struct vhf_data_4 *vhf4;
-
-	vhf4 = (struct vhf_data_4 *)get_vhf_byte(4);
-	if (!vhf4)
-		return;
-	vhf4->WRTP = (flag) ? 1 : 0;
-}
-
-void set_current_state(int s) {
-	uint8_t *vhf_device_activity;
-
-	current_state = s;
-
-	vhf_device_activity = (uint8_t *)get_vhf_byte(6); /* Get DT device activity */
-	if (!vhf_device_activity)
-		return;
-
-	/* Now translate the 'mhVTL' state into DT values */
-	switch (s) {
-	case MHVTL_STATE_UNLOADED:
-		*vhf_device_activity = 0;
-		break;
-	case MHVTL_STATE_LOADING:
-		*vhf_device_activity = 2;
-		break;
-	case MHVTL_STATE_LOADING_CLEAN:
-		*vhf_device_activity = 1;
-		break;
-	case MHVTL_STATE_LOADING_WORM:
-		*vhf_device_activity = 2;
-		break;
-	case MHVTL_STATE_LOADED:
-		*vhf_device_activity = 0;
-		break;
-	case MHVTL_STATE_LOADED_IDLE:
-		*vhf_device_activity = 0;
-		break;
-	case MHVTL_STATE_LOAD_FAILED:
-		*vhf_device_activity = 0;
-		set_lp_11_macc(0); /* MAM Accessible - False */
-		break;
-	case MHVTL_STATE_REWIND:
-		*vhf_device_activity = 0x8;
-		break;
-	case MHVTL_STATE_POSITIONING:
-		*vhf_device_activity = 0x7;
-		break;
-	case MHVTL_STATE_LOCATE:
-		*vhf_device_activity = 0x7;
-		break;
-	case MHVTL_STATE_READING:
-		*vhf_device_activity = 0x5;
-		break;
-	case MHVTL_STATE_WRITING:
-		*vhf_device_activity = 0x6;
-		break;
-	case MHVTL_STATE_UNLOADING:
-		*vhf_device_activity = 0x3;
-		break;
-	case MHVTL_STATE_ERASE:
-		*vhf_device_activity = 0x9;
-		break;
-	case MHVTL_STATE_VERIFY:
-		*vhf_device_activity = 0x4;
-		break;
-	}
-}
-
 static void finish_mount(int sig) {
 	MHVTL_DBG(3, "+++ Trace - Received signal %d +++", sig);
 	if (get_tape_load_status() == TAPE_LOADING)
