+++ usr/ssc.c
@@ -73,87 +73,6 @@ void memset_ssc_buf(struct scsi_cmd *cmd, uint64_t alloc_len) {
 	memset(buf, 0, min((int)alloc_len, lu_priv->bufsize));
 }
 
-/* FIXME: Add VHF log page stuff here */
-int get_tape_load_status(void) {
-	return lu_ssc.load_status;
-}
-
-void set_tape_load_status(int s) {
-	struct vhf_data_5 *vhf5;
-
-	lu_ssc.load_status = s;
-
-	vhf5 = (struct vhf_data_5 *)get_vhf_byte(5);
-
-	if (vhf5) {
-		switch (s) {
-		case TAPE_UNLOADED:
-			vhf5->INXTN	  = 1; /* In transition */
-			vhf5->MSTD	  = 0; /* Medium seated */
-			vhf5->MTHRD	  = 0; /* Medium threaded */
-			vhf5->MOUNTED = 0; /* Medium mounted */
-			vhf5->MPRSNT  = 0; /* Medium Present */
-			vhf5->RAA	  = 1; /* Robotic access allowed */
-			vhf5->INXTN	  = 0; /* Completed updates */
-			set_lp_11_macc(0); /* MAM Accessible */
-			break;
-		case TAPE_LOADED:
-			vhf5->INXTN	  = 1; /* In transition */
-			vhf5->MSTD	  = 1; /* Medium seated */
-			vhf5->MTHRD	  = 1; /* Medium threaded */
-			vhf5->MOUNTED = 1; /* Medium mounted */
-			vhf5->MPRSNT  = 1; /* Medium Present */
-			vhf5->RAA	  = 1; /* Robotic access allowed */
-			vhf5->INXTN	  = 0; /* Completed updates */
-			set_lp_11_macc(1); /* MAM Accessible */
-			break;
-		case TAPE_LOADING:
-			vhf5->INXTN	  = 1; /* In transition */
-			vhf5->MSTD	  = 1; /* Medium seated */
-			vhf5->MTHRD	  = 0; /* Medium threaded */
-			vhf5->MOUNTED = 0; /* Medium mounted */
-			vhf5->MPRSNT  = 1; /* Medium Present */
-			vhf5->RAA	  = 1; /* Robotic access allowed */
-			vhf5->INXTN	  = 0; /* Completed updates */
-			set_lp_11_macc(0); /* MAM Accessible */
-			break;
-		}
-	}
-}
-
-/* Update MAM Accessible bit in LogPage 0x11 */
-void set_lp_11_macc(int flag) {
-	struct vhf_data_4 *vhf4;
-
-	vhf4 = (struct vhf_data_4 *)get_vhf_byte(4);
-	if (!vhf4)
-		return;
-	vhf4->MACC = (flag) ? 1 : 0;
-}
-
-void set_lp11_medium_present(int flag) {
-	struct vhf_data_5 *vhf5;
-
-	vhf5 = (struct vhf_data_5 *)get_vhf_byte(5);
-	if (!vhf5)
-		return;
-	vhf5->MPRSNT = (flag) ? 1 : 0;
-
-	if (!flag) {		   /* Clearing bit - also set state to unloaded */
-		set_lp_11_macc(0); /* MAM Accessible */
-		set_current_state(MHVTL_STATE_UNLOADED);
-	}
-}
-
-void set_lp11_compression(int flag) {
-	struct vhf_data_4 *vhf4;
-
-	vhf4 = (struct vhf_data_4 *)get_vhf_byte(4);
-	if (!vhf4)
-		return;
-	vhf4->CMPR = (flag) ? 1 : 0;
-}
-
 uint8_t ssc_allow_overwrite(struct scsi_cmd *cmd) {
 	declare_ssc_vars;
 
@@ -1848,52 +1767,6 @@ uint8_t ssc_pr_in(struct scsi_cmd *cmd) {
 		return SAM_STAT_RESERVATION_CONFLICT;
 	else
 		return resp_spc_pri(cdb, dbuf_p);
-}
-
-static void update_tape_usage(struct TapeUsage *b) {
-	uint64_t datasets = count_filemarks(-1);
-	uint64_t load_count;
-
-	/* if we have more than 1 filemark,
-	 * most apps write 2 filemarks to flag EOD
-	 * So, lets subtract one from the filemark count to
-	 * present a more accurate 'Data Set' count
-	 */
-	if (datasets > 1)
-		datasets--;
-
-	load_count = get_unaligned_be64(lu_ssc.mamp->LoadCount);
-	put_unaligned_be32(load_count, &b->volumeMounts);
-
-	put_unaligned_be64(datasets, &b->volumeDatasetsWritten);
-}
-
-static void update_seq_access_counters(struct seqAccessDevice *sa) {
-	put_unaligned_be64(lu_ssc.bytesWritten_I,
-					   &sa->writeDataB4Compression);
-	put_unaligned_be64(lu_ssc.bytesWritten_M,
-					   &sa->writeDataAfCompression);
-	put_unaligned_be64(lu_ssc.bytesRead_M,
-					   &sa->readDataB4Compression);
-	put_unaligned_be64(lu_ssc.bytesRead_I,
-					   &sa->readDataAfCompression);
-
-	/* Values in MBytes */
-	if (get_tape_load_status() == TAPE_LOADED) {
-		put_unaligned_be32(lu_ssc.max_capacity >> 20,
-						   &sa->capacity_bop_eod);
-		put_unaligned_be32(lu_ssc.early_warning_position >> 20,
-						   &sa->capacity_bop_ew);
-		put_unaligned_be32(lu_ssc.early_warning_sz >> 20,
-						   &sa->capacity_ew_leop);
-		put_unaligned_be32(current_tape_offset() >> 20,
-						   &sa->capacity_bop_curr);
-	} else {
-		put_unaligned_be32(0xffffffff, &sa->capacity_bop_eod);
-		put_unaligned_be32(0xffffffff, &sa->capacity_bop_ew);
-		put_unaligned_be32(0xffffffff, &sa->capacity_ew_leop);
-		put_unaligned_be32(0xffffffff, &sa->capacity_bop_curr);
-	}
 }
 
 uint8_t ssc_log_sense(struct scsi_cmd *cmd) {
